+++ bin/copy-plugin.sh
@@ -10,12 +10,14 @@ fi
 SOURCE_DIR="$1"
 
 # Define the files and directories to zip
-FILES_TO_COPY="Block Controller etc External Gateway Helper Logger Model Plugin Setup Test view CHANGELOG.md LICENSE README.md registration.php composer.json setup-compile.sh"
+FILES_TO_COPY="Block Controller etc External Gateway Helper Logger Model Plugin Setup Test view CHANGELOG.md LICENSE README.md registration.php composer.json"
 
 if [ "$SOURCE_DIR" = "devilbox" ]; then
     DESTINATION="../devilbox/data/www/magento/magento2/app/code/Xendit/M2Invoice"
 elif [ "$SOURCE_DIR" = "demosites" ]; then
     DESTINATION="../public_html/app/code/Xendit/M2Invoice"
+elif [ "$SOURCE_DIR" = "local-74" ]; then
+    DESTINATION="./docker-magento/magento2/app/code/Xendit/M2Invoice"
 else
     echo "Invalid source directory"
     exit 1
@@ -28,7 +30,10 @@ cp -r $FILES_TO_COPY $DESTINATION || mkdir -p $DESTINATION && cp -r $FILES_TO_CO
 
 # Check if the copy was successful
 if [ $? -eq 0 ]; then
-    echo "Files copied successfully, please run 'bash setup-compile.sh' to compile the module"
+    echo "Files copied successfully, please run 'bash setup-compile.sh' under folder \"app\" to compile the module"
+    if [ "$SOURCE_DIR" = "local-74" ]; then
+        docker exec web bash -c "cd /app && bash setup-compile.sh"
+    fi
 else
     echo "Failed to copy files"
     proxit 1